In "Air_Purify_Sys" v0.1, encoder accuracy suffered from "modeLogic()" delays, addressed in RITOS_Encoder_v2 with 10ms detection intervals.
"Air_Purify_Sys" is the main folder for the project, the LCD could be interact with the encoder in v0.1. However, the detection ability of the encoder was not that good, and it is supposed caused by the longtime calculation for the function "modeLogic()". And in the RITOS_Encoder_v2, a new idea was used to make the every detecting turn 10 ms, hopefully it could help with the accuracy in Air_Purify_Sys_v0.1